+++ irs-api/src/main/java/org/eclipse/tractusx/irs/aaswrapper/job/delegate/BpdmDelegate.java
@@ -54,52 +54,54 @@ public class BpdmDelegate extends AbstractDelegate {
     @Override
     public ItemContainer process(final ItemContainer.ItemContainerBuilder itemContainerBuilder,
             final JobParameter jobData, final AASTransferProcess aasTransferProcess, final String itemId) {
-        final RequestMetric requestMetric = new RequestMetric();
-        requestMetric.setType(RequestMetric.RequestType.BPDM);
 
-        try {
-            itemContainerBuilder.build()
-                                .getBpns()
-                                .forEach(bpn -> lookupBPN(itemContainerBuilder, itemId, bpn, jobData.isLookupBPNs(), requestMetric));
-        } catch (final RestClientException e) {
-            log.info("Business Partner endpoint could not be retrieved for Item: {}. Creating Tombstone.", itemId);
-            requestMetric.incrementFailed();
-            itemContainerBuilder.tombstone(Tombstone.from(itemId, null, e, retryCount, ProcessStep.BPDM_REQUEST))
-                                .metric(requestMetric);
+        if (jobData.isLookupBPNs()) {
+            log.debug("BPN Lookup enabled, collecting BPN information");
+
+            final RequestMetric requestMetric = new RequestMetric();
+            requestMetric.setType(RequestMetric.RequestType.BPDM);
+            itemContainerBuilder.metric(requestMetric);
+
+            try {
+                itemContainerBuilder.build()
+                                    .getBpns()
+                                    .forEach(bpn -> lookupBPN(itemContainerBuilder, itemId, bpn,
+                                            requestMetric));
+            } catch (final RestClientException e) {
+                log.info("Business Partner endpoint could not be retrieved for Item: {}. Creating Tombstone.", itemId);
+                requestMetric.incrementFailed();
+                itemContainerBuilder.tombstone(Tombstone.from(itemId, null, e, retryCount, ProcessStep.BPDM_REQUEST));
+            }
+        } else {
+            log.debug("BPN lookup disabled, no BPN information will be collected.");
         }
 
         return next(itemContainerBuilder, jobData, aasTransferProcess, itemId);
     }
 
     private void lookupBPN(final ItemContainer.ItemContainerBuilder itemContainerBuilder, final String itemId,
-            final Bpn bpn, final boolean bpnLookupEnabled, final RequestMetric metric) {
-        if (bpnLookupEnabled) {
-            log.debug("BPN Lookup enabled, collecting BPN information");
-            bpdmFacade.findManufacturerName(bpn.getManufacturerId()).ifPresentOrElse(name -> {
-                if (BPN_RGX.matcher(bpn.getManufacturerId() + bpn.getManufacturerName()).find()) {
-                    bpn.updateManufacturerName(name);
-                    metric.incrementCompleted();
-                    itemContainerBuilder.metric(metric);
-                } else {
-                    final String message = String.format("BPN: \"%s\" for CatenaXId: %s is not valid.",
-                            bpn.getManufacturerId() + bpn.getManufacturerName(), itemId);
-                    log.warn(message);
-                    metric.incrementFailed();
-                    itemContainerBuilder.tombstone(
-                            Tombstone.from(itemId, null, new BpdmDelegateProcessingException(message), 0,
-                                    ProcessStep.BPDM_VALIDATION)).metric(metric);
-                }
-            }, () -> {
-                final String message = String.format("BPN not exist for given ManufacturerId: %s and for CatenaXId: %s.",
-                        bpn.getManufacturerId(), itemId);
+            final Bpn bpn, final RequestMetric metric) {
+        bpdmFacade.findManufacturerName(bpn.getManufacturerId()).ifPresentOrElse(name -> {
+            if (BPN_RGX.matcher(bpn.getManufacturerId() + bpn.getManufacturerName()).find()) {
+                bpn.updateManufacturerName(name);
+                metric.incrementCompleted();
+            } else {
+                final String message = String.format("BPN: \"%s\" for CatenaXId: %s is not valid.",
+                        bpn.getManufacturerId() + bpn.getManufacturerName(), itemId);
                 log.warn(message);
                 metric.incrementFailed();
-                itemContainerBuilder.tombstone(Tombstone.from(itemId, null, new BpdmDelegateProcessingException(message), 0,
-                        ProcessStep.BPDM_REQUEST)).metric(metric);
-            });
-        } else {
-            log.debug("BPN lookup disabled, no BPN information will be collected.");
-        }
+                itemContainerBuilder.tombstone(
+                        Tombstone.from(itemId, null, new BpdmDelegateProcessingException(message), 0,
+                                ProcessStep.BPDM_VALIDATION));
+            }
+        }, () -> {
+            final String message = String.format("BPN not exist for given ManufacturerId: %s and for CatenaXId: %s.",
+                    bpn.getManufacturerId(), itemId);
+            log.warn(message);
+            metric.incrementFailed();
+            itemContainerBuilder.tombstone(Tombstone.from(itemId, null, new BpdmDelegateProcessingException(message), 0,
+                    ProcessStep.BPDM_REQUEST));
+        });
     }
 
 }