+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/EdcCallbackController.java
@@ -26,8 +26,8 @@ import io.swagger.v3.oas.annotations.Hidden;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.edc.spi.types.domain.edr.EndpointDataReference;
-import org.eclipse.tractusx.irs.common.ApiConstants;
-import org.eclipse.tractusx.irs.common.Masker;
+import org.eclipse.tractusx.irs.data.StringMapper;
+import org.eclipse.tractusx.irs.edc.client.util.Masker;
 import org.eclipse.tractusx.irs.edc.client.configuration.JsonLdConfiguration;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -38,15 +38,15 @@ import org.springframework.web.bind.annotation.RestController;
  * Endpoint used by the EDC ControlPlane to provide the endpoint data reference.
  */
 @Slf4j
-@RestController
-@RequestMapping(ApiConstants.API_PREFIX_INTERNAL)
+@RestController("irsEdcClientEdcCallbackController")
+@RequestMapping("${irs-edc-client.callback.mapping:internal/endpoint-data-reference}")
 @Hidden
 @RequiredArgsConstructor
 public class EdcCallbackController {
 
     private final EndpointDataReferenceStorage storage;
 
-    @PostMapping("/endpoint-data-reference")
+    @PostMapping
     public void receiveEdcCallback(final @RequestBody EndpointDataReference dataReference) {
         log.debug("Received EndpointDataReference: {}", StringMapper.mapToString(dataReference));
         log.debug("Received EndpointDataReference with ID {} and endpoint {}", dataReference.getId(),